+++ lamb/log/constants.py
@@ -7,6 +7,10 @@ __all__ = [
     "LAMB_LOG_FORMAT_CELERY_TASK_SIMPLE",
     "LAMB_LOG_FORMAT_CELERY_MAIN_PREFIXNO",
     "LAMB_LOG_FORMAT_CELERY_TASK_PREFIXNO",
+    "LAMB_LOG_FORMAT_GUNICORN_SIMPLE",
+    "LAMB_LOG_FORMAT_GUNICORN_PREFIXNO",
+    "LAMB_LOG_FORMAT_GUNICORN_VERBOSE",
+    "LAMB_LOG_FORMAT_GUNICORN_VERBOSE_PREFIXNO",
 ]
 
 # api
@@ -16,6 +20,14 @@ LAMB_LOG_FORMAT_SIMPLE = (
 LAMB_LOG_FORMAT_PREFIXNO = "[%(asctime)s, xray=%(xray)s, user_id=%(app_user_id)s: %(levelname)9s, %(prefixno)04d, %(asctime)s-%(prefixno)04d] %(message)s"  # noqa: E501
 LAMB_LOG_FORMAT_VERBOSE = "[%(asctime)s, xray=%(xray)s, user_id=%(app_user_id)s: %(levelname)9s, %(name)s:%(filename)s:%(lineno)4d]  %(message)s"  # noqa: E501
 LAMB_LOG_FORMAT_VERBOSE_PREFIXNO = "[%(asctime)s, xray=%(xray)s, user_id=%(app_user_id)s: %(levelname)9s, %(name)s:%(filename)s:%(lineno)4d, %(prefixno)04d, %(asctime)s-%(prefixno)04d] %(message)s"  # noqa: E501
+
+# gunicorn
+LAMB_LOG_FORMAT_GUNICORN_SIMPLE = "[%(asctime)s, xray=None, user_id=None: %(levelname)9s] %(message)s"  # noqa: E501
+LAMB_LOG_FORMAT_GUNICORN_PREFIXNO = (
+    "[%(asctime)s, xray=None, user_id=None: %(levelname)9s, 0001, %(asctime)s-0001] %(message)s"  # noqa: E501
+)
+LAMB_LOG_FORMAT_GUNICORN_VERBOSE = "[%(asctime)s, xray=None, user_id=None: %(levelname)9s, %(name)s:%(filename)s:%(lineno)4d]  %(message)s"  # noqa: E501
+LAMB_LOG_FORMAT_GUNICORN_VERBOSE_PREFIXNO = "[%(asctime)s, xray=None, user_id=None: %(levelname)9s, %(name)s:%(filename)s:%(lineno)4d, 0001, %(asctime)s-0001] %(message)s"  # noqa: E501
 
 # celery
 LAMB_LOG_FORMAT_CELERY_MAIN_SIMPLE = "[%(asctime)s: %(levelname)9s/%(processName)s] %(message)s"  # noqa: E501